introduced kafka for schedules updated spring boot 3.3.7 to 3.4.2 New endpoint for effective task conf Sending scheduled tasks to kafka Lock heartbeat and release scheduled tasks from task executor Dataset configuration:   new optional column ordering   new constraints declaration Added jdbc demo task processor Added Spark demo task processor

+++ lakehouse-task-executor-svc/src/main/java/org/lakehouse/taskexecutor/service/TaskLockHeartBeat.java
@@ -11,12 +11,12 @@ public class TaskLockHeartBeat implements Runnable{
 	private final Logger logger = LoggerFactory.getLogger(this.getClass());
 	private final SchedulerRestClientApi schedulerRestClientApi;
 	private final TaskExecutionHeartBeatDTO taskExecutionHeartBeatDTO;
-	private final Integer heartBeatIntervalMs;
+	private final Long heartBeatIntervalMs;
 	private boolean exit;
 	
 	public TaskLockHeartBeat(
 			SchedulerRestClientApi schedulerRestClientApi,
-			Integer heartBeatIntervalMs,
+			Long heartBeatIntervalMs,
 			TaskExecutionHeartBeatDTO taskExecutionHeartBeatDTO ) {
 		this.schedulerRestClientApi = schedulerRestClientApi;
 		this.heartBeatIntervalMs = heartBeatIntervalMs;
